+++ PyIRTAM/lib.py
@@ -25,8 +25,11 @@ import os
 import PyIRI
 import PyIRI.main_library as ml
 
+from PyIRTAM import coeff
 
-def IRTAM_density(dtime, alon, alat, modip, TOV, coeff_dir, irtam_dir):
+
+def IRTAM_density(dtime, alon, alat, modip, TOV, irtam_dir='',
+                  use_subdirs=True):
     """Output ionospheric parameters from daily set of IRTAM coefficients.
 
     Parameters
@@ -45,10 +48,13 @@ def IRTAM_density(dtime, alon, alat, modip, TOV, coeff_dir, irtam_dir):
         Modified dip angle in degrees.
     TOV : float
         Time of Validity for IRtAM coefficients (use 24 if unknown).
-    coeff_dir : str
-        Place where model coefficients are located.
     irtam_dir : str
-        Place where IRTAM coefficients are located (user provided).
+        Directory for IRTAM coefficients, or '' to use package directory.
+        (default='')
+    use_subdirs : bool
+        If True, adds YYYY/MMDD subdirectories to the filename path, if False
+        assumes that the entire path to the coefficient directory is provided
+        by `irtam_dir` (default=True)
 
     Returns
     -------
@@ -89,7 +95,7 @@ def IRTAM_density(dtime, alon, alat, modip, TOV, coeff_dir, irtam_dir):
 
     # ------------------------------------------------------------------------
     # Read IRTAM coefficients and form matrix U
-    F_B0, F_B1, F_f0F2, F_hmF2 = IRTAM_read_coeff(dtime, irtam_dir)
+    F_B0, F_B1, F_f0F2, F_hmF2 = IRTAM_read_coeff(dtime, irtam_dir, use_subdirs)
 
     # ------------------------------------------------------------------------
     # Multiply matrices (F_D U)F_G
@@ -97,8 +103,8 @@ def IRTAM_density(dtime, alon, alat, modip, TOV, coeff_dir, irtam_dir):
                                      F_B0, F_B1, F_f0F2, F_hmF2)
 
     # Limit B1 values (same as in IRI) to prevent ugly thin profiles:
-    B1 = np.clip(B1, 1., 6.)
-    B0 = np.clip(B0, 1., 350.)
+    B1 = np.clip(B1, 1.0, 6.0)
+    B0 = np.clip(B0, 1.0, 350.0)
     # ------------------------------------------------------------------------
     # Add all parameters to dictionaries:
     F2 = {'B0': B0,
@@ -203,7 +209,7 @@ def IRTAM_diurnal_functions(time_array, TOV):
     return D
 
 
-def IRTAM_read_coeff(dtime, coeff_dir):
+def IRTAM_read_coeff(dtime, coeff_dir='', use_subdirs=True):
     """Read coefficients from IRTAM.
 
     Parameters
@@ -211,17 +217,22 @@ def IRTAM_read_coeff(dtime, coeff_dir):
     dtime : int
         Month.
     coeff_dir : str
-        Place where the coefficient files are.
+        Directory for IRTAM coefficients, or '' to use package directory.
+        (default='')
+    use_subdirs : bool
+        If True, adds YYYY/MMDD subdirectories to the filename path, if False
+        assumes that the entire path to the coefficient directory is provided
+        by `irtam_dir` (default=True)
 
     Returns
     -------
-    F_B0_IRTAM : array-like
+    b0_irtam : array-like
         IRTAM coefficients for B0 thickness.
-    F_B1_IRTAM : array-like
+    b1_irtam : array-like
         IRTAM coefficients for B1 thickness.
-    F_fof2_IRTAM : array-like
+    fof2_irtam : array-like
         IRTAM coefficients for F2 frequency.
-    F_hmf2_IRTAM : array-like
+    hmf2_irtam : array-like
         IRTAM coefficients for F2 peak height.
 
     Notes
@@ -248,43 +259,33 @@ def IRTAM_read_coeff(dtime, coeff_dir):
     in ionospheric mapping 476 by numerical methods.
 
     """
-    # IRTAM coefficients:
-    # the file has first 988 numbers in same format as CCIR
-    # followed by 76 coefficients for the additional diurnal term
-    time_str = ''.join([dtime.strftime('%Y%m%d'), '_', dtime.strftime('%H%M%S'),
-                        '.ASC'])
-
+    # IRTAM coefficient files have the first 988 numbers in same format as CCIR
+    # followed by 76 coefficients for the additional diurnal term. Load the
+    # parameter data into arrays from the various files.
+    #
     # B0
-    file_B0 = os.path.join(coeff_dir, dtime.strftime('%Y'),
-                           dtime.strftime('%m%d'),
-                           ('IRTAM_B0in_COEFFS_' + time_str))
-
-    F_B0 = IRTAM_read_files(file_B0)
+    file_b0 = coeff.get_irtam_param_filename(dtime, 'B0', irtam_dir=coeff_dir,
+                                             use_subdirs=use_subdirs)
+    b0_irtam = IRTAM_read_files(file_b0)
 
     # B1
-    file_B1 = os.path.join(coeff_dir, dtime.strftime('%Y'),
-                           dtime.strftime('%m%d'),
-                           ('IRTAM_B1in_COEFFS_' + time_str))
-
-    F_B1 = IRTAM_read_files(file_B1)
+    file_b1 = coeff.get_irtam_param_filename(dtime, 'B1', irtam_dir=coeff_dir,
+                                             use_subdirs=use_subdirs)
+    b1_irtam = IRTAM_read_files(file_b1)
 
     # f0F2
-    file_f0F2 = os.path.join(coeff_dir, dtime.strftime('%Y'),
-                             dtime.strftime('%m%d'),
-                             ('IRTAM_foF2_COEFFS_' + time_str))
-
-    F_f0F2 = IRTAM_read_files(file_f0F2)
+    file_f0f2 = coeff.get_irtam_param_filename(dtime, 'foF2',
+                                               irtam_dir=coeff_dir,
+                                               use_subdirs=use_subdirs)
+    f0f2_irtam = IRTAM_read_files(file_f0f2)
 
     # hmF2
-    file_hmF2 = os.path.join(coeff_dir, dtime.strftime('%Y'),
-                             dtime.strftime('%m%d'),
-                             ('IRTAM_hmF2_COEFFS_' + time_str))
+    file_hmf2 = coeff.get_irtam_param_filename(dtime, 'hmF2',
+                                               irtam_dir=coeff_dir,
+                                               use_subdirs=use_subdirs)
+    hmf2_irtam = IRTAM_read_files(file_hmf2)
 
-    F_hmF2 = IRTAM_read_files(file_hmF2)
-
-    output = (F_B0, F_B1, F_f0F2, F_hmF2)
-
-    return output
+    return b0_irtam, b1_irtam, f0f2_irtam, hmf2_irtam
 
 
 def IRTAM_read_files(filename):
@@ -841,7 +842,7 @@ def Ramakrishnan_Rawer_function(NmF2, hmF2, B0, B1, h):
 
 
 def call_IRTAM_PyIRI(aUT, dtime, alon, alat, aalt, f2, f1, e_peak, es_peak,
-                     modip, TOV, coeff_dir, irtam_dir):
+                     modip, TOV, irtam_dir='', use_subdirs=True):
     """Update parameters and build EDP for IRTAM for one time frame.
 
     Parameters
@@ -891,10 +892,13 @@ def call_IRTAM_PyIRI(aUT, dtime, alon, alat, aalt, f2, f1, e_peak, es_peak,
         Modified dip angle in degrees.
     TOV : float
         Time of Validity in decimal hours. Use 24 if not known.
-    coeff_dir : str
-        Direction of IRI coefficients.
     irtam_dir : str
-        Direction of IRTAM coefficients.
+        Directory for IRTAM coefficients, or '' to use package directory.
+        (default='')
+    use_subdirs : bool
+        If True, adds YYYY/MMDD subdirectories to the filename path, if False
+        assumes that the entire path to the coefficient directory is provided
+        by `irtam_dir` (default=True)
 
     Returns
     -------
@@ -937,13 +941,8 @@ def call_IRTAM_PyIRI(aUT, dtime, alon, alat, aalt, f2, f1, e_peak, es_peak,
     it = np.where(aUT == UT)[0]
 
     # Find IRTAM parameters
-    IRTAM_f2 = IRTAM_density(dtime,
-                             alon,
-                             alat,
-                             modip,
-                             TOV,
-                             coeff_dir,
-                             irtam_dir)
+    IRTAM_f2 = IRTAM_density(dtime, alon, alat, modip, TOV, irtam_dir,
+                             use_subdirs)
 
     # Create empty arrays with needed shape for 1 time frame
     # to fill with updated values
@@ -1010,7 +1009,8 @@ def call_IRTAM_PyIRI(aUT, dtime, alon, alat, aalt, f2, f1, e_peak, es_peak,
     return F2_result, F1_result, E_result, Es_result, EDP_result
 
 
-def run_PyIRTAM(year, month, day, aUT, alon, alat, aalt, F107, irtam_dir):
+def run_PyIRTAM(year, month, day, aUT, alon, alat, aalt, F107, irtam_dir='',
+                use_subdirs=True):
     """Update parameters and build EDP for IRTAM for one time frame.
 
     Parameters
@@ -1033,7 +1033,12 @@ def run_PyIRTAM(year, month, day, aUT, alon, alat, aalt, F107, irtam_dir):
     F107 : float
         User provided F10.7 solar flux index in SFU.
     irtam_dir : str
-        Place where IRTAM coefficients are on user's local computer.
+        Directory with IRTAM coefficients, or '' to use package directory.
+        (default='')
+    use_subdirs : bool
+        If True, adds YYYY/MMDD subdirectories to the filename path, if False
+        assumes that the entire path to the coefficient directory is provided
+        by `irtam_dir` (default=True)
 
     Returns
     -------
@@ -1149,19 +1154,11 @@ def run_PyIRTAM(year, month, day, aUT, alon, alat, aalt, F107, irtam_dir):
         dtime = dt.datetime(year, month, day, hour, minute, 0)
 
         # Call PyIRTAM:
-        F2, F1, E, Es, EDP = call_IRTAM_PyIRI(aUT,
-                                              dtime,
-                                              alon,
-                                              alat,
-                                              aalt,
-                                              f2_b,
-                                              f1_b,
-                                              e_b,
-                                              es_b,
-                                              mag['modip'],
-                                              aUT[it],
-                                              PyIRI.coeff_dir,
-                                              irtam_dir)
+        F2, F1, E, Es, EDP = call_IRTAM_PyIRI(aUT, dtime, alon, alat, aalt,
+                                              f2_b, f1_b, e_b, es_b,
+                                              mag['modip'], aUT[it],
+                                              PyIRI.coeff_dir, irtam_dir,
+                                              use_subdirs=use_subdirs)
         # Save results.
         if it == 0:
             for key in F2: